add likes, comments count to headers
