+++ components/Feed/Comments/CommentHeader.tsx
@@ -1,11 +1,13 @@
 import Timestamp from '../../Utils/Timestamp';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import needsHook from '../../../hooks/needsHook';
 import { avatarURL, postCardClass } from '../../../styles/feed';
 import bull from '../../Utils/Bullet';
 import PostOptionsDropdown from '../Post/PostOptionsDropdown';
 import { db } from '../../../firebase';
 import { Avatar } from '@mui/material';
+import { useCollection } from 'react-firebase-hooks/firestore';
+import { UilCornerUpLeftAlt, UilThumbsUp } from '@iconscout/react-unicons';
 
 type CommentHeaderProps = {
     postId: string,
@@ -18,7 +20,44 @@ type CommentHeaderProps = {
 };
 
 const CommentHeader = ({postId, commentId, userImage, name, authorUid, email, timestamp}: CommentHeaderProps) => {
+    const [numLikes, setNumLikes] = useState(0);
+
+    // Use useEffect to bind on document loading the
+    // function that will set the number of likes on
+    // each change of the DB (triggered by onSnapshot)
+    useEffect(() => {
+        db.collection("posts")
+        .doc(postId)
+        .collection("comments")
+        .doc(commentId)
+        .onSnapshot((snapshot) => {
+            // Get the likes map
+            const likesMap = snapshot.data().likes;
     
+            // Count the entries that are True
+            let ctr = 0;
+            for (const [key, value] of Object.entries(likesMap)) {
+            if (value) {
+                ctr += 1;
+            }
+            }
+            setNumLikes(ctr);
+        });
+    }, []);
+
+    // Track number of comments
+    const [repliesSnapshot] = useCollection(
+        db.collection("posts")
+        .doc(postId)
+        .collection("comments")
+        .doc(commentId)
+        .collection("replies")
+    );
+
+    const getNumLikes = () => {
+        return numLikes;
+    };
+
     // Deletes a post
     const deletePost = () => {
         // OPEN A MODAL OR ASK THE USER IF HE/SHE IS SURE TO DELETE THE POST
@@ -70,8 +109,20 @@ const CommentHeader = ({postId, commentId, userImage, name, authorUid, email, ti
                     </div>
 
                     <div className={postCardClass.leftMobileRowTwo}>
+                        {/* Number of replies */}
                         <span className={postCardClass.bullSpan}>{bull}</span> 
+                        <p className={postCardClass.commentsP}>{`${repliesSnapshot ? repliesSnapshot.docs.length : "0"}`}
+                            <span className={postCardClass.commentsSpan}> Replies</span>
+                            <span className={postCardClass.commentsIconSpan}><UilCornerUpLeftAlt size={14}/></span>
+                        </p>
+                        {/* Number of likes */}
+                        {bull}
+                        <p className={postCardClass.commentsP}>{`${getNumLikes() > 0 ? getNumLikes() : "0"}`}
+                            <span className={postCardClass.commentsSpan}> Likes</span>
+                            <span className={postCardClass.commentsIconSpan}><UilThumbsUp size={14}/></span>
+                        </p>
                         {/* Time stamp */}
+                        {bull}
                         <Timestamp timestamp={timestamp} />
                     </div>
                 </div>
