+++ components/Feed/Post/PostHeader.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import needsHook from '../../../hooks/needsHook';
 import { auth, db, storage } from '../../../firebase';
 import { useAuthState } from 'react-firebase-hooks/auth';
@@ -8,6 +8,7 @@ import { postCardClass, avatarURL } from '../../../styles/feed';
 import bull from '../../Utils/Bullet';
 import Timestamp from '../../Utils/Timestamp';
 import { useCollection } from 'react-firebase-hooks/firestore';
+import { UilComment, UilThumbsUp } from '@iconscout/react-unicons'
 
 type PostHeaderProps = {
     id: string,
@@ -20,6 +21,28 @@ type PostHeaderProps = {
 
 const PostHeader = ({id, userImage, name, authorUid, email, timestamp}: PostHeaderProps) => {
     const [user] = useAuthState(auth);
+    const [numLikes, setNumLikes] = useState(0);
+    
+    // Use useEffect to bind on document loading the
+    // function that will set the number of likes on
+    // each change of the DB (triggered by onSnapshot)
+    useEffect(() => {
+        db.collection("posts")
+        .doc(id)
+        .onSnapshot((snapshot) => {
+            // Get the likes map
+            const likesMap = snapshot.data().likes;
+    
+            // Count the entries that are True
+            let ctr = 0;
+            for (const [key, value] of Object.entries(likesMap)) {
+            if (value) {
+                ctr += 1;
+            }
+            }
+            setNumLikes(ctr);
+        });
+    }, []);
 
     // Track number of comments
     const [commentsSnapshot] = useCollection(
@@ -57,6 +80,10 @@ const PostHeader = ({id, userImage, name, authorUid, email, timestamp}: PostHead
         return true
     }
 
+    const getNumLikes = () => {
+        return numLikes;
+    };
+
     return (
         <div className={postCardClass.header}>
             {/* Left content */}
@@ -80,13 +107,19 @@ const PostHeader = ({id, userImage, name, authorUid, email, timestamp}: PostHead
                         <span className={postCardClass.bullSpan}>{bull}</span> 
                         <p className={postCardClass.commentsP}>{`${commentsSnapshot ? commentsSnapshot.docs.length : "0"}`}
                             <span className={postCardClass.commentsSpan}> Comments</span>
+                            <span className={postCardClass.commentsIconSpan}><UilComment size={14}/></span>
+                        </p>
+                        {/* Number of likes */}
+                        {bull} 
+                        <p className={postCardClass.commentsP}>{`${getNumLikes() > 0 ? getNumLikes() : "0"}`}
+                            <span className={postCardClass.commentsSpan}> Likes</span>
+                            <span className={postCardClass.commentsIconSpan}><UilThumbsUp size={14}/></span>
                         </p>
                         {/* TODO: interpolate post category below */}
                         {bull} <p className={postCardClass.categoryP}>Education</p>
                         {bull}
                         {/* Time stamp */}
                         <Timestamp timestamp={timestamp} />
-                        {/* TODO: Figure out where to put comments count */}
                     </div>
                 </div>
             </div>
