+++ components/Feed/Replies/ReplyHeader.tsx
@@ -1,11 +1,12 @@
 import Timestamp from '../../Utils/Timestamp';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import needsHook from '../../../hooks/needsHook';
 import { avatarURL, postCardClass, replyHeaderClass } from '../../../styles/feed';
 import bull from '../../Utils/Bullet';
 import PostOptionsDropdown from '../Post/PostOptionsDropdown';
 import { db } from '../../../firebase';
 import { Avatar } from '@mui/material';
+import { UilThumbsUp } from "@iconscout/react-unicons";
 
 type ReplyHeaderProps = {
     postId: string | string[] | undefined,
@@ -29,7 +30,37 @@ const ReplyHeader: React.FC<ReplyHeaderProps> = (
         email,
         timestamp
     }) => {
+    const [numLikes, setNumLikes] = useState(0);
+
+    // Use useEffect to bind on document loading the
+    // function that will set the number of likes on
+    // each change of the DB (triggered by onSnapshot)
+    useEffect(() => {
+        db.collection("posts")
+        .doc(postId)
+        .collection("comments")
+        .doc(commentId)
+        .collection("replies")
+        .doc(replyId)
+        .onSnapshot((snapshot) => {
+            // Get the likes map
+            const likesMap = snapshot.data().likes;
     
+            // Count the entries that are True
+            let ctr = 0;
+            for (const [key, value] of Object.entries(likesMap)) {
+            if (value) {
+                ctr += 1;
+            }
+            }
+            setNumLikes(ctr);
+        });
+    }, []);
+    
+    const getNumLikes = () => {
+        return numLikes;
+    };
+
     // Deletes a post
     const deletePost = () => {
         // OPEN A MODAL OR ASK THE USER IF HE/SHE IS SURE TO DELETE THE POST
@@ -74,7 +105,13 @@ const ReplyHeader: React.FC<ReplyHeaderProps> = (
                     </div>
 
                     <div className={postCardClass.leftMobileRowTwo}>
-                        <span className={postCardClass.bullSpan}>{bull}</span> 
+                        {/* Number of likes */}
+                        <span className={postCardClass.bullSpan}>{bull}</span>  
+                        <p className={postCardClass.commentsP}>{`${getNumLikes() > 0 ? getNumLikes() : "0"}`}
+                            <span className={postCardClass.commentsSpan}> Likes</span>
+                            <span className={postCardClass.commentsIconSpan}><UilThumbsUp size={14}/></span>
+                        </p>
+                        {bull}
                         {/* Time stamp */}
                         <Timestamp timestamp={timestamp} />
                     </div>
