+++ styles/feed.js
@@ -49,8 +49,9 @@ export const postCardClass = {
   headerLeft: "flex w-11/12 items-center",
   headerRight: "flex w-1/12 justify-end",
   bullSpan: "hidden lg:flex",
-  commentsP: "inline-flex",
+  commentsP: "inline-flex items-center",
   commentsSpan: "hidden md:flex md:ml-1",
+  commentsIconSpan: "flex ml-1 md:hidden",
   categoryP: "font-bold text-primary",
   leftMobileRowOne: "inline-flex items-center",
   leftMobileRowTwo: "inline-flex items-center ml-sm md:ml  text-xs md:text-sm",
